adorn up a tabyl output

Creates total, percetage and N to a tabyl format.

adorn_2x2(tabyl)

adorn_ame(tabyl, where = "col", denominator = "row",
  position = "front", na.rm = TRUE)

adorn_tpn(tabyl, where = "col", denominator = "row",
  position = "front", na.rm = TRUE)

Arguments

tabyl
tabyl input

where
(janitor::adorn_totals) one of "row", "col", or c("row", "col")

denominator
(janitor::adorn_percentages) the direction to use for calculating percentages. One of "row", "col", or "all"

position
(from janitor::adorn_ns) should the N go in the "front", or in the "rear", of the percentage?

na.rm
should missing values (including NaN) be omitted from the calculations?

Value

tabyl object with horizontal

Functions

adorn_2x2: DEPRECATED Creates the usual 2 by 2 output to a tabyl format.

adorn_ame: Creates total, percetage and N to a tabyl format.

adorn_tpn: Creates total, percetage and N to a tabyl format. (same as adorn_ame, but with a formal name)

Examples

library(tidyverse)

#> -- Attaching packages -------------------------------------------- tidyverse 1.2.1 --

#> v ggplot2 3.2.1     v purrr   0.3.3
#> v tibble  2.1.3     v dplyr   0.8.4
#> v tidyr   1.0.0     v stringr 1.4.0
#> v readr   1.3.1     v forcats 0.5.0

#> Warning: package 'forcats' was built under R version 3.6.3

#> -- Conflicts ----------------------------------------------- tidyverse_conflicts() --
#> x dplyr::filter() masks stats::filter()
#> x dplyr::lag()    masks stats::lag()

library(janitor)

#> 
#> Attaching package: 'janitor'

#> The following objects are masked from 'package:stats':
#> 
#>     chisq.test, fisher.test

library(avallecam)

mtcars %>%
  tabyl(am,cyl) %>%
  adorn_ame()

#>  am         4         6          8       Total
#>   0 3 (15.8%) 4 (21.1%) 12 (63.2%) 19 (100.0%)
#>   1 8 (61.5%) 3 (23.1%)  2 (15.4%) 13 (100.0%)

mtcars %>%
  tabyl(am,cyl) %>%
  adorn_ame(where = "row")

#>     am          4         6          8
#>      0  3 (15.8%) 4 (21.1%) 12 (63.2%)
#>      1  8 (61.5%) 3 (23.1%)  2 (15.4%)
#>  Total 11 (34.4%) 7 (21.9%) 14 (43.8%)

mtcars %>%
  tabyl(am,cyl) %>%
  adorn_ame(where = "row",denominator = "col")

#>     am           4          6           8
#>      0  3  (27.3%) 4  (57.1%) 12  (85.7%)
#>      1  8  (72.7%) 3  (42.9%)  2  (14.3%)
#>  Total 11 (100.0%) 7 (100.0%) 14 (100.0%)

mtcars %>%
  tabyl(am,cyl) %>%
  adorn_ame(where = c("row","col"),denominator = "col")

#>     am           4          6           8       Total
#>      0  3  (27.3%) 4  (57.1%) 12  (85.7%) 19  (59.4%)
#>      1  8  (72.7%) 3  (42.9%)  2  (14.3%) 13  (40.6%)
#>  Total 11 (100.0%) 7 (100.0%) 14 (100.0%) 32 (100.0%)

mtcars %>%
  tabyl(am,cyl) %>%
  adorn_ame(where = c("row","col"),denominator = "row")

#>     am          4         6          8       Total
#>      0  3 (15.8%) 4 (21.1%) 12 (63.2%) 19 (100.0%)
#>      1  8 (61.5%) 3 (23.1%)  2 (15.4%) 13 (100.0%)
#>  Total 11 (34.4%) 7 (21.9%) 14 (43.8%) 32 (100.0%)