tidy up incidence outputs

functions to get tibble outputs from incidence package

get_info_tidy(incidence_fit, what)

tidy_incidence(incidence_fit)

glance_incidence(incidence_fit)

Arguments

incidence_fit
incidence_fit or incidence_fit_list object

what
character string of parameter of interest

Value

incidence tibbles!

Functions

get_info_tidy: generates a tibble of incidence::get_info()

tidy_incidence: generates a complete summary tibble from incidence fit paramteter estimates

glance_incidence: generates a complete summary tibble from incidence fit model performance

Examples

# example outbreak --------------------------------------------------------

library(outbreaks) #sample data
library(incidence) #core functions
# library(avallecam) #improvements
dat <- ebola_sim$linelist$date_of_onset
i.7 <- incidence(dat, interval=7)
# plot(i.7)
f1 <- fit(i.7[1:20])
# f1
f2 <- fit_optim_split(i.7)
# f2

# get_info_tidy: tibble outputs from fit object ---------------------------

# original
get_info(f1,what = "r")

#> [1] 0.03175771

# new improvement
get_info_tidy(f1,what = "r")

#> # A tibble: 1 x 4
#>   mark  parameter key       value
#>   <chr> <chr>     <chr>     <dbl>
#> 1 1     r         estimate 0.0318

# for incidence::fit_optim_split outputs use .$fit

# original
get_info(f2$fit, what = "r")

#>      before       after 
#>  0.02982209 -0.01016191 

#improvement
get_info_tidy(f2$fit, what = "r")

#> # A tibble: 2 x 4
#>   mark   parameter key        value
#>   <chr>  <chr>     <chr>      <dbl>
#> 1 before r         estimate  0.0298
#> 2 after  r         estimate -0.0102

# try more examples:

# get_info(f2$fit, what = "r.conf")
# get_info_tidy(f2$fit, what = "r.conf")
# get_info_tidy(f2$fit, what = "doubling")
# get_info_tidy(f2$fit, what = "doubling.conf")
# get_info_tidy(f2$fit, what = "halving")
# get_info_tidy(f2$fit, what = "halving.conf")

# tidy_incidence: all key parameters in tibble ----------------------------

tidy_incidence(incidence_fit = f1)

#> # A tibble: 2 x 5
#>   mark  parameter estimate conf_lower conf_upper
#>   <chr> <chr>        <dbl>      <dbl>      <dbl>
#> 1 1     r           0.0318     0.0260     0.0376
#> 2 1     doubling   21.8       18.5       26.7   

tidy_incidence(incidence_fit = f2$fit)

#> # A tibble: 4 x 5
#>   mark   parameter estimate conf_lower conf_upper
#>   <chr>  <chr>        <dbl>      <dbl>      <dbl>
#> 1 before r           0.0298     0.0261    0.0336 
#> 2 after  r          -0.0102    -0.0110   -0.00930
#> 3 before doubling   23.2       20.7      26.6    
#> 4 after  halving    68.2       62.9      74.5    

# glance_incidence: check model performance -------------------------------

glance_incidence(incidence_fit = f1)

#> # A tibble: 1 x 11
#>   r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>       <dbl>         <dbl> <dbl>     <dbl>    <dbl> <int>  <dbl> <dbl> <dbl>
#> 1     0.880         0.874 0.498      133. 9.81e-10     2  -13.4  32.8  35.7
#> # ... with 2 more variables: deviance <dbl>, df.residual <int>

glance_incidence(incidence_fit = f2$fit)

#> # A tibble: 2 x 12
#>   names r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC
#>   <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <int>  <dbl> <dbl>
#> 1 befo~     0.922         0.919 0.455      273. 2.95e-14     2  -14.8  35.5
#> 2 after     0.951         0.949 0.155      578. 3.72e-21     2   15.4 -24.8
#> # ... with 3 more variables: BIC <dbl>, deviance <dbl>, df.residual <int>