Crear suma de horas por rango de horarios

Problema: Dado el reporte de horas de salida y retorno a vivienda, calcular la suma de horas dentro de cada rango establecido

sum_range_h(dt, lim_sal_h, lim_ret_h, obs_sal_h, obs_ret_h, prefix, suffix)

Arguments

dt
data frame

lim_sal_h
hora limite de salida

lim_ret_h
hora limite de retorno

obs_sal_h
hora observada de salida

obs_ret_h
hora observada de retorno

prefix
prefijo de la nueva variable a crear

suffix
sufijo de la nueva variable a crear

Value

nueva columna con suma de horas por rango

Examples

#crear base
library(tidyverse)
test <- tibble(obs_sal_hx=c(07,04,07,04,24,11,NA,20,22,04),
               obs_ret_hx=c(10,08,13,13,21,NA,NA,09,04,22))
test

#> # A tibble: 10 x 2
#>    obs_sal_hx obs_ret_hx
#>         <dbl>      <dbl>
#>  1          7         10
#>  2          4          8
#>  3          7         13
#>  4          4         13
#>  5         24         21
#>  6         11         NA
#>  7         NA         NA
#>  8         20          9
#>  9         22          4
#> 10          4         22

#Ejemplo:
#Crear 04 rangos de 06 horas cada uno
#sumar la cantidad de horas en dicho rango
#- 06.00 - 11.59 (06, 07, 08, 09, 10, 11)
#- 12.00 - 17.59 (12, 13, 14, 15, 16, 17)
#- 18.00 - 23.59 (18, 19, 20, 21, 22, 23)
#- 00.00 - 05.59 (24, 01, 02, 03, 04, 05)

#crear suma por rango
test %>%
sum_range_h(lim_sal_h = 6,
            lim_ret_h = 11,
            obs_sal_h = obs_sal_hx,
            obs_ret_h = obs_ret_hx,prefix = "mpu_",suffix = "_h1") %>%
sum_range_h(lim_sal_h = 12,
           lim_ret_h = 17,
           obs_sal_h = obs_sal_hx,
           obs_ret_h = obs_ret_hx,prefix = "mpu_",suffix = "_h1") %>%
sum_range_h(lim_sal_h = 18,
           lim_ret_h = 23,
           obs_sal_h = obs_sal_hx,
           obs_ret_h = obs_ret_hx,prefix = "mpu_",suffix = "_h1") %>%
sum_range_h(lim_sal_h = 00,
           lim_ret_h = 05,
           obs_sal_h = obs_sal_hx,
           obs_ret_h = obs_ret_hx,prefix = "mpu_",suffix = "_h1") %>%

#rangos extra
sum_range_h(lim_sal_h = 19,
           lim_ret_h = 24,
           obs_sal_h = obs_sal_hx,
           obs_ret_h = obs_ret_hx,prefix = "mpu_",suffix = "_h1") %>%
sum_range_h(lim_sal_h = 17,
           lim_ret_h = 22,
           obs_sal_h = obs_sal_hx,
           obs_ret_h = obs_ret_hx,prefix = "mpu_",suffix = "_h1")

#> # A tibble: 10 x 8
#>    obs_sal_hx obs_ret_hx mpu_0611_h1 mpu_1217_h1 mpu_1823_h1 mpu_0005_h1
#>         <dbl>      <dbl>       <dbl>       <dbl>       <dbl>       <dbl>
#>  1          7         10           4          NA          NA          NA
#>  2          4          8           3          NA          NA           2
#>  3          7         13           5           2          NA          NA
#>  4          4         13           6           2          NA           2
#>  5         24         21          NA          NA           4          NA
#>  6         11         NA          NA          NA          NA          NA
#>  7         NA         NA          NA          NA          NA          NA
#>  8         20          9           4          NA           4          NA
#>  9         22          4          NA          NA           2           5
#> 10          4         22           6           6           5           2
#> # ... with 2 more variables: mpu_1924_h1 <dbl>, mpu_1722_h1 <dbl>